EMRClusterWithHbase/requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_emr
